move keras exceptions to submodule
